feat: repackage DTO (task/tag), add builder to dto
separated task and tag DTOs by repackaging, and added builder, AllArgsConstructor, NoArgsConstructor annotations to needed DTOs.
