+++ src/main/java/com/wwme/wwme/task/domain/DTO/tagReceiveDTO/CreateTagReceiveDTO.java
@@ -1,0 +1,13 @@
+package com.wwme.wwme.task.domain.DTO.tagReceiveDTO;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class CreateTagReceiveDTO {
+    private long group_id;
+    private String tag_name;
+}
