+++ src/main/java/com/wwme/wwme/task/domain/DTO/taskReceiveDTO/CreateTaskReceiveDTO.java
@@ -1,25 +1,20 @@
-package com.wwme.wwme.task.domain.DTO.sendDTO;
+package com.wwme.wwme.task.domain.DTO.taskReceiveDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class TaskListReadbyUserSendDTO {
-    private Long task_id;
+public class CreateTaskReceiveDTO {
     private String task_name;
     private LocalDateTime start_time;
     private LocalDateTime end_time;
     private String task_type;
     private Long tag_id;
-    private String tag_name;
     private Long group_id;
-    private String group_color;
-    private Boolean is_done_personal;
-    private Boolean is_done_total;
+    private Long todo_user_id;
 }
