+++ src/main/java/com/wwme/wwme/task/domain/DTO/taskReceiveDTO/UpdateTaskReceiveDTO.java
@@ -1,4 +1,4 @@
-package com.wwme.wwme.task.domain.DTO.sendDTO;
+package com.wwme.wwme.task.domain.DTO.taskReceiveDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -9,16 +9,14 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class ReadTaskListByUserSendDTO {
+public class UpdateTaskReceiveDTO {
     private Long task_id;
     private String task_name;
     private LocalDateTime start_time;
     private LocalDateTime end_time;
     private String task_type;
     private Long tag_id;
-    private String tag_name;
     private Long group_id;
-    private String group_color;
-    private Boolean is_done_personal;
     private Boolean is_done_total;
+    private Boolean is_done_personal;
 }
