+++ src/main/java/com/wwme/wwme/task/domain/DTO/taskSendDTO/ReadTaskListByUserSendDTO.java
@@ -1,22 +1,24 @@
-package com.wwme.wwme.task.domain.DTO.receiveDTO;
+package com.wwme.wwme.task.domain.DTO.taskSendDTO;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
-@RequiredArgsConstructor
-public class UpdateTaskReceiveDTO {
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class ReadTaskListByUserSendDTO {
     private Long task_id;
     private String task_name;
     private LocalDateTime start_time;
     private LocalDateTime end_time;
     private String task_type;
     private Long tag_id;
+    private String tag_name;
     private Long group_id;
-    private Boolean is_done_total;
+    private String group_color;
     private Boolean is_done_personal;
+    private Boolean is_done_total;
 }
