+++ src/main/java/com/wwme/wwme/task/domain/DTO/taskSendDTO/TaskListReadbyUserSendDTO.java
@@ -1,21 +1,24 @@
-package com.wwme.wwme.task.domain.DTO.receiveDTO;
+package com.wwme.wwme.task.domain.DTO.taskSendDTO;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class CreateTaskReceiveDTO {
+public class TaskListReadbyUserSendDTO {
+    private Long task_id;
     private String task_name;
     private LocalDateTime start_time;
     private LocalDateTime end_time;
     private String task_type;
     private Long tag_id;
+    private String tag_name;
     private Long group_id;
-    private Long todo_user_id;
+    private String group_color;
+    private Boolean is_done_personal;
+    private Boolean is_done_total;
 }
